# knowledge/GPT_OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: AuthiChain Verification API
  description: Blockchain-powered product authentication. Verify products, generate artistic QRON codes, and check certificates.
  version: 2.0.0
servers:
  - url: https://api.authichain.com/gpt
    description: AuthiChain Gateway (Production)

paths:
  /verify:
    post:
      operationId: verifyProduct
      summary: Verify product authenticity
      description: Run the 5-Agent AI Consensus Protocol to verify a product's authenticity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  description: Product ID or serial number
                imageUrl:
                  type: string
                  description: URL of product image for visual verification
                barcode:
                  type: string
                  description: Barcode or QR code data
                metadata:
                  type: object
                  additionalProperties:
                    type: string
                  description: Additional product metadata
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  trustScore:
                    type: number
                  verdict:
                    type: string
                    enum: [authentic, suspicious, counterfeit]
                  agents:
                    type: object
                  certificate:
                    type: string

  /qr/generate:
    post:
      operationId: generateQronCode
      summary: Generate an artistic QRON code
      description: Create an AI-generated artistic QR code in one of 11 style modes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
              properties:
                data:
                  type: string
                  description: Data to encode (URL, product ID, or text)
                style:
                  type: string
                  enum:
                    - cosmic_nebula
                    - cyberpunk
                    - watercolor
                    - holographic_mosaic
                    - teal_pulse
                    - medieval_fantasy
                    - anime_manga
                    - architectural
                    - echo_qr
                    - video_qr
                    - phantom_qr
                  description: Art style for the QRON
                prompt:
                  type: string
                  description: Custom art prompt for QR background
      responses:
        "200":
          description: Generated QRON
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
                  qrData:
                    type: string
                  style:
                    type: string

  /certificates/verify:
    get:
      operationId: checkCertificate
      summary: Look up a verification certificate
      description: Retrieve certificate details and provenance chain by certificate number.
      parameters:
        - name: certNumber
          in: query
          required: true
          schema:
            type: string
          description: Certificate number (e.g., AC-2026-XXXX)
      responses:
        "200":
          description: Certificate details
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificateNumber:
                    type: string
                  productName:
                    type: string
                  verifiedAt:
                    type: string
                  trustScore:
                    type: number
                  provenanceChain:
                    type: array
                    items:
                      type: object

  /cannabis/verify:
    post:
      operationId: verifyCannabisStrain
      summary: Verify a cannabis strain
      description: Validate strain genetics, lab results, and METRC compliance via StrainChain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - strainName
              properties:
                strainName:
                  type: string
                  description: Name of the cannabis strain
                batchId:
                  type: string
                  description: Batch or lot number
                labCertHash:
                  type: string
                  description: SHA-256 hash of the lab certificate
                thcPercent:
                  type: number
                  description: Claimed THC percentage
                cbdPercent:
                  type: number
                  description: Claimed CBD percentage
      responses:
        "200":
          description: Strain verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  strainName:
                    type: string
                  rarityScore:
                    type: number
                  labCertValid:
                    type: boolean
                  metrcCompliant:
                    type: boolean

  /trust-score:
    post:
      operationId: computeTrustScore
      summary: Compute composite trust score
      description: Calculate a trust score from multiple verification signals.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - qrDecodePass
                - blockchainCertExists
              properties:
                qrDecodePass:
                  type: boolean
                  description: Whether QR code decoded successfully
                blockchainCertExists:
                  type: boolean
                  description: Whether blockchain certificate exists
                nfcMatch:
                  type: boolean
                  description: Whether NFC chip data matches
                visualMatch:
                  type: number
                  description: Visual similarity score (0-100)
                geoFenceOk:
                  type: boolean
                  description: Whether scan is within geo-fence
      responses:
        "200":
          description: Trust score result
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: number
                  verdict:
                    type: string
                  breakdown:
                    type: object